instance_manager: Move out Error to its own module
